[brockit] Make Rebase v2 official
This PR deletes the old rebase v1 one code, and makes the new v2 code
the official. This code has been tested now for quite a few weeks, and
it is stable enough to be made official.

Resolves https://github.com/brave/brave-browser/issues/55466
