+++ tools/cr/rebase_test.py
@@ -3,31 +3,1040 @@
 # This Source Code Form is subject to the terms of the Mozilla Public
 # License, v. 2.0. If a copy of the MPL was not distributed with this file,
 # You can obtain one at https://mozilla.org/MPL/2.0/.
-"""Tests for `brockit.Rebase`.
+"""Tests for `rebase`.
 
-This file is organised in two layers:
+This file is organised in three layers:
 
-* `RebaseStaticHelpersTest` covers the four `@staticmethod` rebase-todo-file
-  transformers as pure file-in/file-out unit tests. No git involved.
+* Unit tests (`EntryLineParseTest`, `PatternMatchTest`, `RewritePlanTest`,
+  `MessageWriterTest`) exercise the pure file-in / file-out transforms in the
+  `rebase` module with no git involved.
 
-* `RebaseExecuteTest` covers `Rebase.execute()` end-to-end against a real git
-  repository tree built by `FakeChromiumRepo`. The editor callbacks
+* `RebaseRecommitHelperTest` covers the `brockit.Rebase.recommit_in_rebase_plan`
+  todo-file transformer as a pure file-in / file-out unit test.
+
+* `RebaseExecuteTest` covers `brockit.Rebase.execute()` end-to-end against a
+  real git tree built by `FakeChromiumRepo`. The editor callbacks
   (`GIT_SEQUENCE_EDITOR`, `GIT_EDITOR`) re-enter `brockit.py`'s `__main__`
   dispatch via subprocess, exactly as in production -- nothing is mocked out.
 """
 
+import os
 import tempfile
 import unittest
-from dataclasses import dataclass
 from pathlib import Path
 from unittest.mock import patch
 
 import brockit
+import rebase
 from test.fake_chromium_repo import FakeChromiumRepo
 
 
-class RebaseStaticHelpersTest(unittest.TestCase):
-    """Tests for the four todo-file transformers on `Rebase`."""
+class EntryLineParseTest(unittest.TestCase):
+    """Tests for `rebase.EntryLine.parse`."""
+
+    def test_parse_simple_pick(self):
+        entry_line = rebase.EntryLine.parse('pick aaa # [cr148] Feature A\n')
+        self.assertEqual(entry_line.command, 'pick')
+        self.assertEqual(entry_line.hash, 'aaa')
+        self.assertEqual(entry_line.message, '[cr148] Feature A')
+        self.assertIsNone(entry_line.subcommand)
+        self.assertIsNone(entry_line.note)
+
+    def test_parse_fixup_dash_c_multi_word_command(self):
+        """`fixup -C <hash>` keeps both tokens as the command."""
+        entry_line = rebase.EntryLine.parse(
+            'fixup -C 2c334294014 # amend! [cr149] Permit cross-host\n')
+        self.assertEqual(entry_line.command, 'fixup -C')
+        self.assertEqual(entry_line.hash, '2c334294014')
+        self.assertEqual(entry_line.message, '[cr149] Permit cross-host')
+        self.assertEqual(entry_line.subcommand, ['amend'])
+
+    def test_parse_reassign_subcommand(self):
+        """`reassign!<hash>!` is captured as two subcommand tokens."""
+        entry_line = rebase.EntryLine.parse(
+            'pick zzz # reassign!c080270dd7e! [cr149] Fix bookmark bar.\n')
+        self.assertEqual(entry_line.subcommand, ['reassign', 'c080270dd7e'])
+        self.assertEqual(entry_line.message, '[cr149] Fix bookmark bar.')
+
+    def test_parse_reassign_subcommand_hash_starting_with_digit(self):
+        """`reassign!<hash>!` where `<hash>` starts with a digit must
+        still capture both tokens. Regression: an earlier regex required
+        every `!`-separated token to start with a letter, which silently
+        dropped digit-leading hashes -- the subcommand collapsed to
+        `['reassign']` and the hash leaked into `message`, causing
+        `rewrite_plan` to orphan-drop the reassign."""
+        entry_line = rebase.EntryLine.parse(
+            'pick 04d3a656bb1 # reassign!859ab9caa74! [cr150][ios] Add '
+            '//brave/ios/browser/svg to visibility for //third_party/expat\n')
+        self.assertEqual(entry_line.subcommand, ['reassign', '859ab9caa74'])
+        self.assertEqual(
+            entry_line.message,
+            '[cr150][ios] Add //brave/ios/browser/svg to visibility for '
+            '//third_party/expat')
+        self.assertEqual(entry_line.reassign_target_hash, '859ab9caa74')
+
+    def test_parse_trailing_empty_note(self):
+        """A trailing ` # empty` marker is split into `note` and stripped
+        out of the comment portion."""
+        entry_line = rebase.EntryLine.parse(
+            'pick zzz # reassign!bbb! [cr148] Feature B # empty\n')
+        self.assertEqual(entry_line.note, 'empty')
+        self.assertEqual(entry_line.message, '[cr148] Feature B')
+        self.assertEqual(entry_line.subcommand, ['reassign', 'bbb'])
+
+    def test_parse_blank_line_returns_none(self):
+        self.assertIsNone(rebase.EntryLine.parse('\n'))
+        self.assertIsNone(rebase.EntryLine.parse('   \n'))
+
+    def test_parse_pure_comment_returns_none(self):
+        self.assertIsNone(rebase.EntryLine.parse('# Rebase plan\n'))
+        self.assertIsNone(rebase.EntryLine.parse('   # indented\n'))
+
+    def test_parse_malformed_line_raises(self):
+        """A non-empty, non-comment line that doesn't fit the
+        `<cmd> <hash> # <msg>` shape raises `EditorRecoverableFailure`
+        so the caller can punt to the user editor."""
+        with self.assertRaises(rebase.EditorRecoverableFailure):
+            rebase.EntryLine.parse('garbage\n')
+        with self.assertRaises(rebase.EditorRecoverableFailure):
+            rebase.EntryLine.parse('pickonly # no hash\n')
+
+    def test_entry_type_classification_and_properties(self):
+        """`entry_type` is filled in by `EntryLine.__init__` during
+        construction;
+        the `is_*` properties are derived from it and stay mutually
+        exclusive."""
+        EntryType = rebase.EntryType
+
+        pinned = rebase.EntryLine.parse(
+            'pick aaa # Update from Chromium 1.0.0.0 to Chromium 1.0.0.1\n')
+        self.assertIs(pinned.entry_type, EntryType.PINNED)
+        self.assertTrue(pinned.is_pinned)
+        self.assertFalse(pinned.is_recyclable)
+        self.assertFalse(pinned.is_reassignment)
+        self.assertEqual(pinned.pinned_group, 'version')
+
+        recyclable = rebase.EntryLine.parse(
+            'pick bbb # Update patches from Chromium 1.0.0.0 to '
+            'Chromium 1.0.0.1.\n')
+        self.assertIs(recyclable.entry_type, EntryType.RECYCLABLE)
+        self.assertTrue(recyclable.is_recyclable)
+
+        reassign = rebase.EntryLine.parse(
+            'pick zzz # reassign!bbb! [cr148] Feature B\n')
+        self.assertIs(reassign.entry_type, EntryType.REASSIGNMENT)
+        self.assertTrue(reassign.is_reassignment)
+
+        regular = rebase.EntryLine.parse('pick ccc # [cr148] Feature A\n')
+        self.assertIs(regular.entry_type, EntryType.DEFAULT)
+        self.assertFalse(regular.is_pinned)
+        self.assertFalse(regular.is_recyclable)
+        self.assertFalse(regular.is_reassignment)
+
+    def test_reassign_target_hash_property(self):
+        """`reassign_target_hash` returns the hash from `reassign!<hash>!`
+        for reassignment commits, `None` when the reassignment has no
+        hash, and raises `NotImplementedError` for non-reassignment
+        commits."""
+        full = rebase.EntryLine.parse(
+            'pick zzz # reassign!c080270dd7e! [cr149] Fix bookmark bar.\n')
+        self.assertEqual(full.reassign_target_hash, 'c080270dd7e')
+
+        no_hash = rebase.EntryLine.parse(
+            'pick zzz # reassign! [cr149] Some subject.\n')
+        self.assertTrue(no_hash.is_reassignment)
+        self.assertIsNone(no_hash.reassign_target_hash)
+
+        regular = rebase.EntryLine.parse('pick aaa # [cr148] Feature A\n')
+        with self.assertRaises(NotImplementedError):
+            _ = regular.reassign_target_hash
+
+    def test_drop_classification_and_target_hash(self):
+        """`drop!<hash>!` commits classify as DROP, expose `is_drop`, and
+        carry their target hash in `drop_target_hash` -- mirroring the
+        reassign equivalents."""
+        EntryType = rebase.EntryType
+
+        drop = rebase.EntryLine.parse(
+            'pick zzz # drop!bbb! [cr148] Feature B\n')
+        self.assertIs(drop.entry_type, EntryType.DROP)
+        self.assertTrue(drop.is_drop)
+        self.assertFalse(drop.is_reassignment)
+        self.assertEqual(drop.drop_target_hash, 'bbb')
+
+        no_hash = rebase.EntryLine.parse(
+            'pick zzz # drop! [cr148] Feature B\n')
+        self.assertTrue(no_hash.is_drop)
+        self.assertIsNone(no_hash.drop_target_hash)
+
+        regular = rebase.EntryLine.parse('pick aaa # [cr148] Feature A\n')
+        with self.assertRaises(NotImplementedError):
+            _ = regular.drop_target_hash
+
+    def test_fixup_of_pinned_classified_as_pinned(self):
+        """A `fixup -C` whose `amend!` subject is pinned still ends up
+        `entry_type=PINNED` with the right `pinned_group`. The
+        autosquash marker is peeled off into `subcommand` by
+        `EntryLine.parse`, so the regular pinned classifier on the bare
+        subject does the work -- no special-case handling needed."""
+        entry_line = rebase.EntryLine.parse(
+            'fixup -C bbb # amend! Update from Chromium 1.0.0.0 to '
+            'Chromium 1.0.0.1\n')
+        self.assertIs(entry_line.entry_type, rebase.EntryType.PINNED)
+        self.assertTrue(entry_line.is_pinned)
+        self.assertEqual(entry_line.pinned_group, 'version')
+
+
+class PatternMatchTest(unittest.TestCase):
+    """Tests for the cr-tag stripper, pinned classifier, and recyclable
+    detector."""
+
+    def test_strip_cr_tag_removes_simple_tag(self):
+        self.assertEqual(rebase._strip_cr_tag('[cr149] Foo'), 'Foo')
+
+    def test_strip_cr_tag_removes_chained_tags(self):
+        """`[cr149][ios] ` collapses into nothing."""
+        self.assertEqual(rebase._strip_cr_tag('[cr149][ios] Foo'), 'Foo')
+        self.assertEqual(rebase._strip_cr_tag('[cr149][ios][extra] Foo'),
+                         'Foo')
+
+    def test_strip_cr_tag_leaves_unrelated_tag(self):
+        """`[unrelated] ` is not a cr-tag; left in place."""
+        self.assertEqual(rebase._strip_cr_tag('[unrelated] Foo'),
+                         '[unrelated] Foo')
+
+    def test_strip_cr_tag_noop_on_untagged(self):
+        self.assertEqual(
+            rebase._strip_cr_tag('Update from Chromium 1.0.0.0 to '
+                                 'Chromium 1.0.0.1'),
+            'Update from Chromium 1.0.0.0 to Chromium 1.0.0.1')
+
+    def test_strip_autosquash_prefixes_handles_stacks(self):
+        """`fixup! ` / `amend! ` / `squash! ` prefixes can stack when a
+        commit fixes up an already-fixup commit. `_strip_autosquash_prefixes`
+        peels off any number of them at the start."""
+        self.assertEqual(
+            rebase._strip_autosquash_prefixes(
+                'fixup! Conflict-resolved patches from Chromium 1.0.0.0 '
+                'to Chromium 1.0.0.1.'),
+            'Conflict-resolved patches from Chromium 1.0.0.0 to '
+            'Chromium 1.0.0.1.')
+        self.assertEqual(
+            rebase._strip_autosquash_prefixes(
+                'fixup! fixup! fixup! Conflict-resolved patches from '
+                'Chromium 1.0.0.0 to Chromium 1.0.0.1.'),
+            'Conflict-resolved patches from Chromium 1.0.0.0 to '
+            'Chromium 1.0.0.1.')
+        # Mixed prefixes also strip cleanly.
+        self.assertEqual(
+            rebase._strip_autosquash_prefixes(
+                'amend! fixup! squash! [cr149] IWYU fixes.'),
+            '[cr149] IWYU fixes.')
+        # Untagged subjects pass through unchanged.
+        self.assertEqual(
+            rebase._strip_autosquash_prefixes('Just a plain subject'),
+            'Just a plain subject')
+
+    def test_pinned_match_survives_stacked_autosquash_prefixes(self):
+        """A pinned subject buried under stacked `fixup!` prefixes still
+        classifies via the autosquash-prefix stripping pre-pass."""
+        self.assertEqual(
+            rebase.get_pinned_group_for_subject(
+                'fixup! fixup! Conflict-resolved patches from Chromium '
+                '1.0.0.0 to Chromium 1.0.0.1.'), 'conflict')
+        self.assertEqual(
+            rebase.get_pinned_group_for_subject('fixup! [cr149] IWYU fixes.'),
+            'iwyu')
+        # Stripping is composable with `[crNNN]` -- the autosquash
+        # strip happens first, then `[crNNN]`.
+        self.assertEqual(
+            rebase.get_pinned_group_for_subject(
+                'amend! [cr149] `gnrt` run for Chromium 149.0.7827.5'), 'gnrt')
+
+    def test_pinned_patterns_match_canonical_subjects(self):
+        """Every pinned pattern matches its canonical subject form, both
+        tagged and untagged where applicable."""
+        cases = [
+            ('version', 'Update from Chromium 1.0.0.0 to Chromium 1.0.0.1'),
+            ('plaster_reruns',
+             'Apply-fixed 🩹 patches from Chromium 1.0.0.0 to '
+             'Chromium 1.0.0.1.'),
+            ('conflict', 'Conflict-resolved patches from Chromium 1.0.0.0 to '
+             'Chromium 1.0.0.1.'),
+            ('gnrt', '[cr149] `gnrt` run for Chromium 149.0.7827.5'),
+            ('iwyu', '[cr149] IWYU fixes.'),
+            ('resource_ids', '[cr149] Bump resource_ids'),
+            ('lit_mangler', '[cr149] Update Lit mangler snapshot'),
+            ('disable_tests', '[cr149] Disable failing upstream tests'),
+        ]
+        for expected_id, subject in cases:
+            with self.subTest(subject=subject):
+                self.assertEqual(rebase.get_pinned_group_for_subject(subject),
+                                 expected_id)
+
+    def test_pinned_iwyu_untagged_form_also_matches(self):
+        """The `[cr*]`-tagged patterns must also match without the tag."""
+        self.assertEqual(rebase.get_pinned_group_for_subject('IWYU fixes.'),
+                         'iwyu')
+        self.assertEqual(
+            rebase.get_pinned_group_for_subject('Bump resource_ids'),
+            'resource_ids')
+
+    def test_pinned_rejects_near_misses(self):
+        """Near-miss subjects should not match any pinned pattern."""
+        self.assertIsNone(
+            rebase.get_pinned_group_for_subject(
+                'Update from Chromium 1.0.0.0 to '
+                'Chromium 1.0.0.1 (revert pending)'))
+        self.assertIsNone(
+            rebase.get_pinned_group_for_subject('[cr149] Feature work'))
+        self.assertIsNone(
+            rebase.get_pinned_group_for_subject('Update from Chromium abc to '
+                                                'Chromium def'))
+
+    def test_pinned_rejects_non_four_segment_versions(self):
+        """Chromium versions must be exactly four dot-separated numeric
+        segments. Three-segment or five-segment shapes are rejected."""
+        self.assertIsNone(
+            rebase.get_pinned_group_for_subject(
+                'Update from Chromium 1.0.0 to '
+                'Chromium 1.0.1'))
+        self.assertIsNone(
+            rebase.get_pinned_group_for_subject(
+                'Update from Chromium 1.0.0.0.0 to '
+                'Chromium 1.0.0.0.1'))
+        self.assertIsNone(
+            rebase.get_pinned_group_for_subject(
+                '`gnrt` run for Chromium 149.0.7827'))
+
+    def test_recyclable_matches_canonical(self):
+        self.assertTrue(
+            rebase.is_recyclable('Update patches from Chromium 1.0.0.0 '
+                                 'to Chromium 1.0.0.1.'))
+        self.assertTrue(
+            rebase.is_recyclable('Updated strings for Chromium 1.0.0.1.'))
+
+    def test_recyclable_rejects_near_misses(self):
+        self.assertFalse(
+            rebase.is_recyclable('Update from Chromium 1.0.0.0 to '
+                                 'Chromium 1.0.0.1'))
+        self.assertFalse(rebase.is_recyclable('[cr149] Feature work'))
+
+
+class RewritePlanTest(unittest.TestCase):
+    """Tests for `rebase.rewrite_plan`.
+
+    Split along the `(pinned_squashed, discard_recyclable)` flag matrix.
+    """
+
+    def setUp(self):
+        tmp = tempfile.TemporaryDirectory()
+        self.addCleanup(tmp.cleanup)
+        self._tmp_root = Path(tmp.name)
+
+    def _todo(self, content: str) -> Path:
+        path = self._tmp_root / 'git-rebase-todo'
+        path.write_text(content)
+        return path
+
+    # ----- pinned_squashed=False, discard_recyclable=False ------------------
+
+    def test_identity_transform_keeps_everything_in_order(self):
+        """With no flags, `rewrite` is a no-op: the file is not parsed
+        and not rewritten. Useful as a fast-path for callers that pass
+        the flags through indiscriminately."""
+        path = self._todo(
+            'pick aaa # Update from Chromium 1.0.0.0 to Chromium 1.0.0.1\n'
+            'pick bbb # [cr148] Feature A\n'
+            'pick ccc # Update patches from Chromium 1.0.0.0 to '
+            'Chromium 1.0.0.1.\n'
+            'pick zzz # reassign!bbb! [cr148] Feature A\n')
+
+        rebase.rewrite_plan(todo_file=path)
+
+        self.assertEqual(
+            path.read_text(),
+            'pick aaa # Update from Chromium 1.0.0.0 to Chromium 1.0.0.1\n'
+            'pick bbb # [cr148] Feature A\n'
+            'pick ccc # Update patches from Chromium 1.0.0.0 to '
+            'Chromium 1.0.0.1.\n'
+            'pick zzz # reassign!bbb! [cr148] Feature A\n')
+
+    def test_identity_drops_comments_and_blanks_when_parsing(self):
+        """When any flag forces a parse pass (here `discard_recyclable`),
+        comments and blank lines are filtered out. With no flags set
+        `rewrite` is a no-op (covered by
+        `test_identity_transform_keeps_everything_in_order`)."""
+        path = self._todo('# Rebase plan generated by git\n'
+                          '\n'
+                          'pick aaa # [cr148] Feature A\n'
+                          '\n')
+
+        rebase.rewrite_plan(todo_file=path, discard_recyclable=True)
+
+        self.assertEqual(path.read_text(), 'pick aaa # [cr148] Feature A\n')
+
+    # ----- pinned_squashed=False, discard_recyclable=True -------------------
+
+    def test_discard_recyclable_drops_only_recyclable_lines(self):
+        """Recyclable lines are dropped; pinned, reassign, and regular
+        commits stay in their arrival order, untouched."""
+        path = self._todo(
+            'pick aaa # Update from Chromium 1.0.0.0 to Chromium 1.0.0.1\n'
+            'pick bbb # [cr148] Feature A\n'
+            'pick ccc # Update patches from Chromium 1.0.0.0 to '
+            'Chromium 1.0.0.1.\n'
+            'pick ddd # Updated strings for Chromium 1.0.0.1.\n'
+            'pick zzz # reassign!bbb! [cr148] Feature A\n')
+
+        rebase.rewrite_plan(todo_file=path, discard_recyclable=True)
+
+        self.assertEqual(
+            path.read_text(),
+            'pick aaa # Update from Chromium 1.0.0.0 to Chromium 1.0.0.1\n'
+            'pick bbb # [cr148] Feature A\n'
+            'pick zzz # reassign!bbb! [cr148] Feature A\n')
+
+    # ----- pinned_squashed=True, discard_recyclable=False -------------------
+
+    def test_squashed_collapses_consecutive_version_bumps(self):
+        path = self._todo(
+            'pick aaa # Update from Chromium 1.0.0.0 to Chromium 1.0.0.1\n'
+            'pick bbb # Update from Chromium 1.0.0.1 to Chromium 1.0.0.2\n'
+            'pick ccc # Update from Chromium 1.0.0.2 to Chromium 1.0.0.3\n')
+
+        rebase.rewrite_plan(todo_file=path, pinned_squashed=True)
+
+        self.assertEqual(
+            path.read_text(),
+            'pick aaa # Update from Chromium 1.0.0.0 to Chromium 1.0.0.1\n'
+            'squash bbb # Update from Chromium 1.0.0.1 to Chromium 1.0.0.2\n'
+            'squash ccc # Update from Chromium 1.0.0.2 to Chromium 1.0.0.3\n')
+
+    def test_squashed_reorders_pinned_groups_by_priority(self):
+        """Pinned groups emit in the priority order:
+        version → plaster_reruns → conflict → gnrt → iwyu → resource_ids
+        → lit_mangler → disable_tests, regardless of arrival order."""
+        path = self._todo(
+            'pick aaa # [cr148] Disable failing upstream tests\n'
+            'pick bbb # [cr148] Update Lit mangler snapshot\n'
+            'pick ccc # [cr148] Bump resource_ids\n'
+            'pick ddd # [cr148] IWYU fixes.\n'
+            'pick eee # [cr148] `gnrt` run for Chromium 1.0.0.1\n'
+            'pick fff # Conflict-resolved patches from Chromium 1.0.0.0 '
+            'to Chromium 1.0.0.1.\n'
+            'pick ggg # Apply-fixed 🩹 patches from Chromium 1.0.0.0 '
+            'to Chromium 1.0.0.1.\n'
+            'pick hhh # Update from Chromium 1.0.0.0 to Chromium 1.0.0.1\n')
+
+        rebase.rewrite_plan(todo_file=path, pinned_squashed=True)
+
+        self.assertEqual(
+            path.read_text(),
+            'pick hhh # Update from Chromium 1.0.0.0 to Chromium 1.0.0.1\n'
+            'pick ggg # Apply-fixed 🩹 patches from Chromium 1.0.0.0 '
+            'to Chromium 1.0.0.1.\n'
+            'pick fff # Conflict-resolved patches from Chromium 1.0.0.0 '
+            'to Chromium 1.0.0.1.\n'
+            'pick eee # [cr148] `gnrt` run for Chromium 1.0.0.1\n'
+            'pick ddd # [cr148] IWYU fixes.\n'
+            'pick ccc # [cr148] Bump resource_ids\n'
+            'pick bbb # [cr148] Update Lit mangler snapshot\n'
+            'pick aaa # [cr148] Disable failing upstream tests\n')
+
+    def test_squashed_keeps_recyclable_lines_in_place(self):
+        """Recyclable lines are NOT grouped or moved with pinned commits.
+        They stay interleaved with regular commits in arrival order."""
+        path = self._todo(
+            'pick aaa # [cr148] Feature A\n'
+            'pick bbb # Update from Chromium 1.0.0.0 to Chromium 1.0.0.1\n'
+            'pick ccc # Update patches from Chromium 1.0.0.0 to '
+            'Chromium 1.0.0.1.\n'
+            'pick ddd # [cr148] Feature B\n')
+
+        rebase.rewrite_plan(todo_file=path, pinned_squashed=True)
+
+        # Pinned (`bbb`) jumps to the top; recyclable (`ccc`) and regulars
+        # (`aaa`, `ddd`) keep their relative order in the tail.
+        self.assertEqual(
+            path.read_text(),
+            'pick bbb # Update from Chromium 1.0.0.0 to Chromium 1.0.0.1\n'
+            'pick aaa # [cr148] Feature A\n'
+            'pick ccc # Update patches from Chromium 1.0.0.0 to '
+            'Chromium 1.0.0.1.\n'
+            'pick ddd # [cr148] Feature B\n')
+
+    def test_squashed_groups_fixup_of_pinned_with_target(self):
+        """A `fixup -C` whose `amend!` subject matches a pinned pattern is
+        grouped with the pinned target, in arrival order."""
+        path = self._todo(
+            'pick aaa # Update from Chromium 1.0.0.0 to Chromium 1.0.0.1\n'
+            'fixup -C bbb # amend! Update from Chromium 1.0.0.0 to '
+            'Chromium 1.0.0.1\n'
+            'pick ccc # [cr148] Feature A\n')
+
+        rebase.rewrite_plan(todo_file=path, pinned_squashed=True)
+
+        # The fixup line is in the version group as `squash`; its original
+        # `fixup -C` command is replaced by `squash` (it is no longer the
+        # first commit in the group).
+        self.assertEqual(
+            path.read_text(),
+            'pick aaa # Update from Chromium 1.0.0.0 to Chromium 1.0.0.1\n'
+            'squash bbb # amend! Update from Chromium 1.0.0.0 to '
+            'Chromium 1.0.0.1\n'
+            'pick ccc # [cr148] Feature A\n')
+
+    def test_squashed_groups_squash_autosquash_of_pinned(self):
+        """A `squash <hash> # squash! <pinned-subject>` line -- distinct
+        from `fixup`/`fixup -C` but the same intent -- is also grouped with
+        the pinned target. A `pick` line with the same autosquash marker
+        is not (it has to be a non-`pick` autosquash command)."""
+        path = self._todo(
+            'pick aaa # [cr148] `gnrt` run for Chromium 1.0.0.1\n'
+            'squash bbb # squash! [cr148] `gnrt` run for Chromium 1.0.0.1\n'
+            'pick ccc # [cr148] Feature A\n'
+            'pick ddd # squash! [cr148] `gnrt` run for Chromium 1.0.0.1\n')
+
+        rebase.rewrite_plan(todo_file=path, pinned_squashed=True)
+
+        # `bbb` (squash autosquash) groups with `aaa`; `ddd` (pick command,
+        # despite the same marker) is treated as a regular pinned commit
+        # and also groups with `aaa` via its message. Both non-first
+        # entries in the group are forced to `squash`.
+        self.assertEqual(
+            path.read_text(),
+            'pick aaa # [cr148] `gnrt` run for Chromium 1.0.0.1\n'
+            'squash bbb # squash! [cr148] `gnrt` run for Chromium 1.0.0.1\n'
+            'squash ddd # squash! [cr148] `gnrt` run for Chromium 1.0.0.1\n'
+            'pick ccc # [cr148] Feature A\n')
+
+    def test_squashed_reassign_matched_by_hash(self):
+        """Reassign with hash `bbb` finds the `pick bbb` line and inserts
+        itself directly above it; target becomes `squash`."""
+        path = self._todo('pick aaa # [cr148] Feature A\n'
+                          'pick bbb # [cr148] Feature B\n'
+                          'pick zzz # reassign!bbb! [cr148] Feature B\n')
+
+        rebase.rewrite_plan(todo_file=path, pinned_squashed=True)
+
+        self.assertEqual(
+            path.read_text(), 'pick aaa # [cr148] Feature A\n'
+            'pick zzz # reassign!bbb! [cr148] Feature B\n'
+            'squash bbb # [cr148] Feature B\n')
+
+    def test_squashed_reassign_target_hash_starts_with_digit(self):
+        """Real-world regression: a `reassign!<hash>!` where the hash
+        starts with a digit (which is the case for ~6/16 short hashes on
+        average) must still be matched to its target. Previously the
+        subcommand regex required each token to start with a letter, so
+        the hash never made it into `subcommand` -- the reassign was
+        silently dropped and the target was left untouched."""
+        path = self._todo(
+            'pick 7606227b5da # [cr150][ios] Add backend promo provider\n'
+            'pick 859ab9caa74 # [cr150][ios] Add //brave/ios/browser/svg '
+            'to visibility for //third_party/expat\n'
+            'pick 04d3a656bb1 # reassign!859ab9caa74! [cr150][ios] Add '
+            '//brave/ios/browser/svg to visibility for '
+            '//third_party/expat # empty\n'
+            'pick 9c160e03412 # [cr150] wip-reassing-bug\n')
+
+        rebase.rewrite_plan(todo_file=path, pinned_squashed=True)
+
+        self.assertEqual(
+            path.read_text(),
+            'pick 7606227b5da # [cr150][ios] Add backend promo provider\n'
+            'pick 04d3a656bb1 # reassign!859ab9caa74! [cr150][ios] Add '
+            '//brave/ios/browser/svg to visibility for '
+            '//third_party/expat # empty\n'
+            'squash 859ab9caa74 # [cr150][ios] Add //brave/ios/browser/svg '
+            'to visibility for //third_party/expat\n'
+            'pick 9c160e03412 # [cr150] wip-reassing-bug\n')
+
+    def test_squashed_reassign_falls_back_to_message_match(self):
+        """Non-matching hash, matching message: the hash lookup misses,
+        so the reassign falls back to matching by commit subject."""
+        path = self._todo('pick aaa # [cr148] Feature A\n'
+                          'pick bbb # [cr148] Feature B\n'
+                          'pick zzz # reassign!xxx! [cr148] Feature B\n')
+
+        rebase.rewrite_plan(todo_file=path, pinned_squashed=True)
+
+        self.assertEqual(
+            path.read_text(), 'pick aaa # [cr148] Feature A\n'
+            'pick zzz # reassign!xxx! [cr148] Feature B\n'
+            'squash bbb # [cr148] Feature B\n')
+
+    def test_squashed_reassign_hash_match_wins_over_message_match(self):
+        """Matching hash, non-matching message: hash lookup has priority.
+        Even though the reassign's message also matches *another* commit
+        in the plan, the hash-matching commit is the one that gets
+        squashed."""
+        path = self._todo('pick aaa # Some completely different subject\n'
+                          'pick bbb # [cr148] Feature B\n'
+                          'pick zzz # reassign!aaa! [cr148] Feature B\n')
+
+        rebase.rewrite_plan(todo_file=path, pinned_squashed=True)
+
+        # `aaa` matches the reassign's hash and is squashed; `bbb`
+        # (whose subject matches the reassign's message) is left
+        # untouched because hash-match took priority.
+        self.assertEqual(
+            path.read_text(), 'pick zzz # reassign!aaa! [cr148] Feature B\n'
+            'squash aaa # Some completely different subject\n'
+            'pick bbb # [cr148] Feature B\n')
+
+    def test_squashed_reassign_without_hash_uses_message_match_only(self):
+        """No-hash-in-subcommand (`reassign! <subject>` with a single
+        `!`): the hash lookup is skipped entirely and the reassign is
+        placed via message-based matching."""
+        path = self._todo('pick aaa # [cr148] Feature A\n'
+                          'pick bbb # [cr148] Feature B\n'
+                          'pick zzz # reassign! [cr148] Feature B\n')
+
+        # Sanity check: the parser sees only one subcommand token, so
+        # `reassign_target_hash` is None and the hash-lookup branch in
+        # `add_reassign_before_target` is skipped.
+        reassign = rebase.EntryLine.parse(
+            'pick zzz # reassign! [cr148] Feature B')
+        self.assertEqual(reassign.subcommand, ['reassign'])
+        self.assertIsNone(reassign.reassign_target_hash)
+
+        rebase.rewrite_plan(todo_file=path, pinned_squashed=True)
+
+        self.assertEqual(
+            path.read_text(), 'pick aaa # [cr148] Feature A\n'
+            'pick zzz # reassign! [cr148] Feature B\n'
+            'squash bbb # [cr148] Feature B\n')
+
+    def test_squashed_reassign_to_pinned_target_is_dropped(self):
+        """Reassignment of a pinned commit is not supported. The reassign
+        line is treated as orphaned and silently dropped; the pinned
+        commit is emitted normally."""
+        path = self._todo(
+            'pick aaa # [cr148] Feature A\n'
+            'pick bbb # Update from Chromium 1.0.0.0 to Chromium 1.0.0.1\n'
+            'pick zzz # reassign!bbb! Update from Chromium 1.0.0.0 to '
+            'Chromium 1.0.0.1\n')
+
+        rebase.rewrite_plan(todo_file=path, pinned_squashed=True)
+
+        self.assertEqual(
+            path.read_text(),
+            'pick bbb # Update from Chromium 1.0.0.0 to Chromium 1.0.0.1\n'
+            'pick aaa # [cr148] Feature A\n')
+
+    def test_squashed_orphan_reassign_dropped(self):
+        """A reassign whose hash and subject both miss is silently
+        dropped."""
+        path = self._todo(
+            'pick aaa # [cr148] Feature A\n'
+            'pick zzz # reassign!xxx! [cr148] Completely unrelated subject\n')
+
+        rebase.rewrite_plan(todo_file=path, pinned_squashed=True)
+
+        self.assertEqual(path.read_text(), 'pick aaa # [cr148] Feature A\n')
+
+    def test_squashed_reassign_with_empty_suffix(self):
+        """The ` # empty` suffix on a reassign must be stripped before the
+        message-based lookup."""
+        path = self._todo(
+            'pick aaa # [cr148] Feature A\n'
+            'pick bbb # [cr148] Feature B\n'
+            'pick zzz # reassign!xxx! [cr148] Feature B # empty\n')
+
+        rebase.rewrite_plan(todo_file=path, pinned_squashed=True)
+
+        self.assertEqual(
+            path.read_text(), 'pick aaa # [cr148] Feature A\n'
+            'pick zzz # reassign!xxx! [cr148] Feature B # empty\n'
+            'squash bbb # [cr148] Feature B\n')
+
+    # ----- pinned_squashed=True, drop! markers ------------------------------
+
+    def test_squashed_drop_removes_target_and_marker(self):
+        """A `drop!<hash>!` removes both its target (matched by hash) and
+        itself from the plan."""
+        path = self._todo('pick aaa # [cr148] Feature A\n'
+                          'pick bbb # [cr148] Feature B\n'
+                          'pick zzz # drop!bbb! [cr148] Feature B\n')
+
+        rebase.rewrite_plan(todo_file=path, pinned_squashed=True)
+
+        self.assertEqual(path.read_text(), 'pick aaa # [cr148] Feature A\n')
+
+    def test_squashed_drop_falls_back_to_message_match(self):
+        """A non-matching hash falls back to dropping the commit whose
+        subject matches the drop's message."""
+        path = self._todo('pick aaa # [cr148] Feature A\n'
+                          'pick bbb # [cr148] Feature B\n'
+                          'pick zzz # drop!xxx! [cr148] Feature B\n')
+
+        rebase.rewrite_plan(todo_file=path, pinned_squashed=True)
+
+        self.assertEqual(path.read_text(), 'pick aaa # [cr148] Feature A\n')
+
+    def test_squashed_orphan_drop_left_untouched(self):
+        """A drop whose hash and subject both miss is silently discarded
+        (its marker line is removed), leaving the rest of the plan alone."""
+        path = self._todo(
+            'pick aaa # [cr148] Feature A\n'
+            'pick zzz # drop!xxx! [cr148] Completely unrelated subject\n')
+
+        rebase.rewrite_plan(todo_file=path, pinned_squashed=True)
+
+        self.assertEqual(path.read_text(), 'pick aaa # [cr148] Feature A\n')
+
+    def test_squashed_drop_with_empty_suffix(self):
+        """The ` # empty` suffix on a drop is stripped before the
+        message-based lookup, just like reassign."""
+        path = self._todo('pick aaa # [cr148] Feature A\n'
+                          'pick bbb # [cr148] Feature B\n'
+                          'pick zzz # drop!xxx! [cr148] Feature B # empty\n')
+
+        rebase.rewrite_plan(todo_file=path, pinned_squashed=True)
+
+        self.assertEqual(path.read_text(), 'pick aaa # [cr148] Feature A\n')
+
+    def test_squashed_drop_not_processed_without_pinned_squashed(self):
+        """Like reassign, drop markers are only acted on under
+        `pinned_squashed`. With only `discard_recyclable`, the marker and its
+        target are left in place."""
+        path = self._todo('pick aaa # [cr148] Feature A\n'
+                          'pick bbb # [cr148] Feature B\n'
+                          'pick zzz # drop!bbb! [cr148] Feature B\n')
+
+        rebase.rewrite_plan(todo_file=path, discard_recyclable=True)
+
+        self.assertEqual(
+            path.read_text(), 'pick aaa # [cr148] Feature A\n'
+            'pick bbb # [cr148] Feature B\n'
+            'pick zzz # drop!bbb! [cr148] Feature B\n')
+
+    # ----- pinned_squashed=True, discard_recyclable=True --------------------
+
+    def test_combined_pinned_grouped_and_recyclable_dropped(self):
+        path = self._todo(
+            'pick aaa # [cr148] Feature A\n'
+            'pick bbb # Update from Chromium 1.0.0.0 to Chromium 1.0.0.1\n'
+            'pick ccc # Update patches from Chromium 1.0.0.0 to '
+            'Chromium 1.0.0.1.\n'
+            'pick ddd # [cr148] Feature B\n')
+
+        rebase.rewrite_plan(todo_file=path,
+                            pinned_squashed=True,
+                            discard_recyclable=True)
+
+        self.assertEqual(
+            path.read_text(),
+            'pick bbb # Update from Chromium 1.0.0.0 to Chromium 1.0.0.1\n'
+            'pick aaa # [cr148] Feature A\n'
+            'pick ddd # [cr148] Feature B\n')
+
+    def test_combined_reassign_orphaned_when_target_is_dropped_recyclable(
+            self):
+        """A reassign targeting a recyclable that gets dropped becomes
+        orphaned in phase B."""
+        path = self._todo(
+            'pick aaa # [cr148] Feature A\n'
+            'pick ccc # Updated strings for Chromium 1.0.0.1.\n'
+            'pick zzz # reassign!ccc! Updated strings for Chromium 1.0.0.1.\n')
+
+        rebase.rewrite_plan(todo_file=path,
+                            pinned_squashed=True,
+                            discard_recyclable=True)
+
+        self.assertEqual(path.read_text(), 'pick aaa # [cr148] Feature A\n')
+
+    # ----- Cross-cutting -----------------------------------------------------
+
+    def test_malformed_line_raises_editor_recoverable(self):
+        """A malformed line surfaces as `EditorRecoverableFailure` --
+        only when `rewrite` actually parses the file (i.e. at least one
+        flag is set). The brockit dispatch catches this and punts to the
+        editor with the failure reason prepended as a `#`-comment."""
+        path = self._todo('pick aaa # [cr148] Feature A\n'
+                          'totally malformed line without hash or hash\n')
+
+        with self.assertRaises(rebase.EditorRecoverableFailure):
+            rebase.rewrite_plan(todo_file=path, discard_recyclable=True)
+
+
+class MessageWriterTest(unittest.TestCase):
+    """Tests for `rebase.MessageWriter`.
+
+    All fixtures use git's verbose squash-editor layout: each
+    `# This is the …` / `# The commit message #N will be skipped:` header
+    is followed by a blank line, then the content. This matches the file
+    git actually writes to `$GIT_EDITOR` during squash rebases.
+    """
+
+    def setUp(self):
+        tmp = tempfile.TemporaryDirectory()
+        self.addCleanup(tmp.cleanup)
+        self._tmp_root = Path(tmp.name)
+
+    def _file(self, content: str) -> Path:
+        path = self._tmp_root / 'COMMIT_EDITMSG'
+        path.write_text(content)
+        return path
+
+    @staticmethod
+    def _messages(writer: 'rebase.MessageWriter') -> list:
+        return [block.full_message for block in writer.blocks]
+
+    def test_pinned_squash_writes_trailing_message(self):
+        """Block 1 is a version-bump subject; block 2 is the next bump.
+        Parse accepts (pinned first block); `rewrite_with_last_message`
+        writes the trailing block's message back to the file."""
+        path = self._file('# This is the 1st commit message:\n'
+                          '\n'
+                          'Update from Chromium 1.0.0.0 to Chromium 1.0.0.1\n'
+                          '\n'
+                          '# This is the commit message #2:\n'
+                          '\n'
+                          'Update from Chromium 1.0.0.1 to Chromium 1.0.0.2\n')
+
+        writer = rebase.MessageWriter.parse(path)
+
+        self.assertEqual(self._messages(writer), [
+            'Update from Chromium 1.0.0.0 to Chromium 1.0.0.1',
+            'Update from Chromium 1.0.0.1 to Chromium 1.0.0.2',
+        ])
+
+        writer.rewrite_with_last_message()
+        self.assertEqual(path.read_text(),
+                         'Update from Chromium 1.0.0.1 to Chromium 1.0.0.2\n')
+
+    def test_reassignment_squash_writes_target_message(self):
+        """A `reassign!` first content line in block 1 is accepted by
+        parse. The reassign block stays in `blocks`, but it's block 2
+        (carrying the original commit's subject and body) that gets
+        written back -- because it's the last block."""
+        path = self._file('# This is the 1st commit message:\n'
+                          '\n'
+                          'reassign!bbb! [cr148] Feature B\n'
+                          '\n'
+                          '# This is the commit message #2:\n'
+                          '\n'
+                          '[cr148] Feature B\n'
+                          '\n'
+                          'Original body line.\n')
+
+        writer = rebase.MessageWriter.parse(path)
+
+        self.assertEqual(self._messages(writer), [
+            'reassign!bbb! [cr148] Feature B',
+            '[cr148] Feature B\n\nOriginal body line.',
+        ])
+
+        writer.rewrite_with_last_message()
+        self.assertEqual(path.read_text(),
+                         '[cr148] Feature B\n\nOriginal body line.\n')
+
+    def test_unclassifiable_first_block_raises(self):
+        """Block 1 whose first content line is neither `reassign!` nor a
+        pinned pattern raises `EditorRecoverableFailure`. The exception
+        message names the offending content line so the caller can
+        surface it to the user."""
+        path = self._file('# This is the 1st commit message:\n'
+                          '\n'
+                          '[cr148] Some regular feature commit\n')
+
+        with self.assertRaises(rebase.EditorRecoverableFailure) as ctx:
+            rebase.MessageWriter.parse(path)
+        self.assertIn('[cr148] Some regular feature commit',
+                      str(ctx.exception))
+
+    def test_parse_failure_wrapped_in_cannot_classify(self):
+        """An empty file (no blocks parse out) also raises
+        `EditorRecoverableFailure`."""
+        path = self._file('')
+        with self.assertRaises(rebase.EditorRecoverableFailure):
+            rebase.MessageWriter.parse(path)
+
+    def test_squash_marker_merges_into_previous_block(self):
+        """A block whose autosquash note is `# squash! ...` is merged
+        into the previous block's `full_message` rather than becoming a
+        new block."""
+        path = self._file(
+            '# This is the 1st commit message:\n'
+            '\n'
+            'Update from Chromium 1.0.0.0 to Chromium 1.0.0.1\n'
+            '\n'
+            '# This is the commit message #2:\n'
+            '\n'
+            '# squash! Update from Chromium 1.0.0.0 to Chromium 1.0.0.1\n'
+            '\n'
+            'Some extra detail\n')
+
+        writer = rebase.MessageWriter.parse(path)
+
+        # Single merged block: block 1's content with the squash content
+        # appended on a new line.
+        self.assertEqual(self._messages(writer), [
+            'Update from Chromium 1.0.0.0 to Chromium 1.0.0.1\n'
+            'Some extra detail',
+        ])
+
+    def test_will_be_skipped_block_contributes_nothing(self):
+        """The skipped-block header is treated as a pure delimiter; its
+        contents never reach `blocks` regardless of marker."""
+        path = self._file(
+            '# This is the 1st commit message:\n'
+            '\n'
+            'Update from Chromium 1.0.0.0 to Chromium 1.0.0.1\n'
+            '\n'
+            '# The commit message #2 will be skipped:\n'
+            '\n'
+            '# fixup! Update from Chromium 1.0.0.0 to Chromium 1.0.0.1\n'
+            '\n'
+            '# This is the commit message #3:\n'
+            '\n'
+            'Update from Chromium 1.0.0.1 to Chromium 1.0.0.2\n')
+
+        writer = rebase.MessageWriter.parse(path)
+
+        self.assertEqual(self._messages(writer), [
+            'Update from Chromium 1.0.0.0 to Chromium 1.0.0.1',
+            'Update from Chromium 1.0.0.1 to Chromium 1.0.0.2',
+        ])
+
+    def test_bare_fixup_block_is_skipped(self):
+        """When the user has stacked `fixup!` commits on top of an
+        already-fixup commit, git emits the marker-only `# fixup!`
+        (or `# fixup! fixup! ...`) inside a regular `# This is the
+        commit message #N:` header with no body. Those blocks carry
+        nothing to merge -- treat them like `will be skipped:` rather
+        than failing the parse."""
+        path = self._file(
+            '# This is a combination of 4 commits.\n'
+            '# This is the 1st commit message:\n'
+            '\n'
+            'Conflict-resolved patches from Chromium 1.0.0.0 to '
+            'Chromium 1.0.0.1.\n'
+            '\n'
+            '# This is the commit message #2:\n'
+            '\n'
+            '# fixup! Conflict-resolved patches from Chromium 1.0.0.0 to '
+            'Chromium 1.0.0.1.\n'
+            '\n'
+            '# This is the commit message #3:\n'
+            '\n'
+            '# fixup! fixup! Conflict-resolved patches from Chromium 1.0.0.0 '
+            'to Chromium 1.0.0.1.\n'
+            '\n'
+            '# This is the commit message #4:\n'
+            '\n'
+            'Conflict-resolved patches from Chromium 1.0.0.2 to '
+            'Chromium 1.0.0.3.\n')
+
+        writer = rebase.MessageWriter.parse(path)
+
+        self.assertEqual(self._messages(writer), [
+            'Conflict-resolved patches from Chromium 1.0.0.0 to '
+            'Chromium 1.0.0.1.',
+            'Conflict-resolved patches from Chromium 1.0.0.2 to '
+            'Chromium 1.0.0.3.',
+        ])
+
+    def test_git_footer_terminates_parsing(self):
+        """Anything after `# Please enter the commit message ...` is
+        ignored, including stale `# This is ...` lines from git's status
+        echo."""
+        path = self._file(
+            '# This is the 1st commit message:\n'
+            '\n'
+            'Update from Chromium 1.0.0.0 to Chromium 1.0.0.1\n'
+            '\n'
+            '# Please enter the commit message for your changes. Lines '
+            'starting\n'
+            '# with \'#\' will be ignored, and an empty message aborts the '
+            'commit.\n'
+            '#\n'
+            '# interactive rebase in progress; onto 5e0f42a4157\n')
+
+        writer = rebase.MessageWriter.parse(path)
+
+        self.assertEqual(self._messages(writer),
+                         ['Update from Chromium 1.0.0.0 to Chromium 1.0.0.1'])
+
+    def test_hand_off_to_editor_invokes_supplied_editor(self):
+        """The caller-supplied `editor` is what gets spawned, via
+        `terminal.run(..., interactive=True)` so the editor inherits the
+        parent tty. `hand_off_to_editor` does not consult any env var
+        of its own -- editor resolution is the caller's job (typically
+        `rebase.get_git_editor`)."""
+        path = self._file('original content\n')
+
+        with patch.object(rebase.terminal, 'run') as mock_run:
+            rebase.hand_off_to_editor(path, reason='', editor='nano')
+        mock_run.assert_called_once_with(['nano', str(path)], interactive=True)
+
+    def test_hand_off_to_editor_splits_multi_token_editor(self):
+        """A multi-token `editor` (e.g. `code --wait`, or a flag with
+        embedded whitespace inside quotes) is split via `shlex.split`
+        before being passed to `terminal.run`, so each token lands in
+        its own argv slot."""
+        path = self._file('original content\n')
+
+        with patch.object(rebase.terminal, 'run') as mock_run:
+            rebase.hand_off_to_editor(path, reason='', editor='code --wait')
+        mock_run.assert_called_once_with(
+            ['code', '--wait', str(path)], interactive=True)
+
+        with patch.object(rebase.terminal, 'run') as mock_run:
+            rebase.hand_off_to_editor(
+                path, reason='', editor='my_editor --feature="with space"')
+        mock_run.assert_called_once_with(
+            ['my_editor', '--feature=with space',
+             str(path)], interactive=True)
+
+    def test_hand_off_to_editor_prepends_reason_as_comment(self):
+        """When `reason` is supplied, every line is prepended to the
+        file as a `#`-comment. Lines that already start with `#` pass
+        through unchanged (no double-`#`)."""
+        path = self._file('original line 1\noriginal line 2\n')
+
+        with patch.object(rebase.terminal, 'run'):
+            rebase.hand_off_to_editor(
+                path,
+                reason='Cannot classify commit\nFix it manually',
+                editor='vim')
+
+        self.assertEqual(
+            path.read_text(),
+            '# B🚀 has been unable to handle this rebase. Investigate '
+            'why, and\n'
+            '# file a bug. Failure reason below.\n'
+            '# Cannot classify commit\n'
+            '# Fix it manually\n'
+            'original line 1\n'
+            'original line 2\n')
+
+    def test_hand_off_to_editor_no_reason_leaves_file_alone(self):
+        """With an empty `reason`, the file content is untouched -- only
+        the editor subprocess is invoked."""
+        path = self._file('original content\n')
+
+        with patch.object(rebase.terminal, 'run'):
+            rebase.hand_off_to_editor(path, reason='', editor='vim')
+
+        self.assertEqual(path.read_text(), 'original content\n')
+
+    def test_get_git_editor_raises_if_env_already_overridden(self):
+        """If the resolved env var holds a brockit `--internal-rebase-*`
+        dispatch, `get_git_editor` is being called too late and would
+        loop on itself; raise `NotImplementedError` instead."""
+        overridden = ('/path/to/vpython3 /path/to/brockit.py '
+                      '--internal-rebase-plan-squash-pinned')
+        with patch.dict(os.environ, {'GIT_EDITOR': overridden}, clear=False):
+            with self.assertRaises(NotImplementedError):
+                rebase.get_git_editor()
+
+
+class RebaseRecommitHelperTest(unittest.TestCase):
+    """Tests for the `brockit.Rebase.recommit_in_rebase_plan` transformer."""
 
     def setUp(self):
         tmp = tempfile.TemporaryDirectory()
@@ -39,67 +1048,6 @@ class RebaseStaticHelpersTest(unittest.TestCase):
         path = self._tmp_root / 'git-rebase-todo'
         path.write_text(content)
         return path
-
-    # ----- discard_regen_changes_from_rebase_plan ----------------------------
-
-    def test_discard_regen_removes_update_patches_line(self):
-        """A line whose comment starts with `Update patches from Chromium ` is
-        dropped; everything else is kept verbatim."""
-        path = self._todo('pick aaa # [cr148] Real feature commit\n'
-                          'pick bbb # Update patches from Chromium 1.0.0.0 '
-                          'to Chromium 1.0.0.1.\n'
-                          'pick ccc # [cr148] Another feature commit\n')
-
-        brockit.Rebase.discard_regen_changes_from_rebase_plan(path)
-
-        self.assertEqual(
-            path.read_text(), 'pick aaa # [cr148] Real feature commit\n'
-            'pick ccc # [cr148] Another feature commit\n')
-
-    def test_discard_regen_removes_updated_strings_line(self):
-        """A line whose comment starts with `Updated strings for Chromium ` is
-        dropped; everything else is kept verbatim."""
-        path = self._todo('pick aaa # [cr148] Real feature commit\n'
-                          'pick bbb # Updated strings for Chromium 1.0.0.0 '
-                          'to Chromium 1.0.0.1.\n'
-                          'pick ccc # [cr148] Another feature commit\n')
-
-        brockit.Rebase.discard_regen_changes_from_rebase_plan(path)
-
-        self.assertEqual(
-            path.read_text(), 'pick aaa # [cr148] Real feature commit\n'
-            'pick ccc # [cr148] Another feature commit\n')
-
-    def test_discard_regen_preserves_comments_blanks_and_other_picks(self):
-        """Comment lines, blank lines, and unrelated picks survive."""
-        path = self._todo('# Rebase plan generated by git\n'
-                          '\n'
-                          'pick aaa # [cr148] Feature A\n'
-                          'pick bbb # Update patches from Chromium 1.0.0.0 '
-                          'to Chromium 1.0.0.1.\n'
-                          'pick ccc # [cr148] Feature B\n'
-                          '\n'
-                          '# Commands:\n')
-
-        brockit.Rebase.discard_regen_changes_from_rebase_plan(path)
-
-        self.assertEqual(
-            path.read_text(), '# Rebase plan generated by git\n'
-            '\n'
-            'pick aaa # [cr148] Feature A\n'
-            'pick ccc # [cr148] Feature B\n'
-            '\n'
-            '# Commands:\n')
-
-    def test_discard_regen_empty_file_is_noop(self):
-        """An empty todo file remains empty."""
-        path = self._todo('')
-
-        brockit.Rebase.discard_regen_changes_from_rebase_plan(path)
-
-        self.assertEqual(path.read_text(), '')
-
-    # ----- recommit_in_rebase_plan -------------------------------------------
 
     def test_recommit_flips_only_first_pick(self):
         """The first occurrence of `pick` becomes `edit`; later `pick`s stay."""
@@ -122,249 +1070,6 @@ class RebaseStaticHelpersTest(unittest.TestCase):
 
         self.assertEqual(path.read_text(), '# Just comments\n\n')
 
-    # ----- squash_minor_bumps_from_rebase_plan -------------------------------
-
-    def test_squash_collapses_consecutive_version_bumps(self):
-        """Repeated `Update from Chromium` commits keep the first as pick and
-        become squash for the rest, preserving their order."""
-        path = self._todo(
-            'pick aaa # Update from Chromium 1.0.0.0 to Chromium 1.0.0.1\n'
-            'pick bbb # Update from Chromium 1.0.0.1 to Chromium 1.0.0.2\n'
-            'pick ccc # Update from Chromium 1.0.0.2 to Chromium 1.0.0.3\n')
-
-        brockit.Rebase.squash_minor_bumps_from_rebase_plan(path)
-
-        self.assertEqual(
-            path.read_text(),
-            'pick aaa # Update from Chromium 1.0.0.0 to Chromium 1.0.0.1\n'
-            'squash bbb # Update from Chromium 1.0.0.1 to Chromium 1.0.0.2\n'
-            'squash ccc # Update from Chromium 1.0.0.2 to Chromium 1.0.0.3\n')
-
-    def test_squash_reorders_categories(self):
-        """The output order is version, plaster reruns, conflict, gnrt, iwyu,
-        then everything else -- regardless of input order."""
-        path = self._todo(
-            'pick aaa # [cr148] Some feature commit\n'
-            'pick bbb # Conflict-resolved patches from Chromium 1.0.0.0 '
-            'to Chromium 1.0.0.1.\n'
-            'pick ccc # Update from Chromium 1.0.0.0 to Chromium 1.0.0.1\n'
-            'pick ddd # [cr148] `gnrt` run for Chromium 1.0.0.1\n'
-            'pick eee # [cr148] IWYU fixes.\n'
-            'pick fff # Apply-fixed 🩹 patches from Chromium 1.0.0.0 '
-            'to Chromium 1.0.0.1.\n')
-
-        brockit.Rebase.squash_minor_bumps_from_rebase_plan(path)
-
-        self.assertEqual(
-            path.read_text(),
-            'pick ccc # Update from Chromium 1.0.0.0 to Chromium 1.0.0.1\n'
-            'pick fff # Apply-fixed 🩹 patches from Chromium 1.0.0.0 '
-            'to Chromium 1.0.0.1.\n'
-            'pick bbb # Conflict-resolved patches from Chromium 1.0.0.0 '
-            'to Chromium 1.0.0.1.\n'
-            'pick ddd # [cr148] `gnrt` run for Chromium 1.0.0.1\n'
-            'pick eee # [cr148] IWYU fixes.\n'
-            'pick aaa # [cr148] Some feature commit\n')
-
-    def test_squash_collapses_plaster_reruns(self):
-        """First `Apply-fixed patches` stays pick; the rest become squash."""
-        path = self._todo(
-            'pick aaa # Apply-fixed 🩹 patches from Chromium 1.0.0.0 '
-            'to Chromium 1.0.0.1.\n'
-            'pick bbb # Apply-fixed 🩹 patches from Chromium 1.0.0.1 '
-            'to Chromium 1.0.0.2.\n')
-
-        brockit.Rebase.squash_minor_bumps_from_rebase_plan(path)
-
-        self.assertEqual(
-            path.read_text(),
-            'pick aaa # Apply-fixed 🩹 patches from Chromium 1.0.0.0 '
-            'to Chromium 1.0.0.1.\n'
-            'squash bbb # Apply-fixed 🩹 patches from Chromium 1.0.0.1 '
-            'to Chromium 1.0.0.2.\n')
-
-    def test_squash_collapses_conflict_resolved(self):
-        """Repeated `Conflict-resolved patches` commits keep first as pick."""
-        path = self._todo(
-            'pick aaa # Conflict-resolved patches from Chromium 1.0.0.0 '
-            'to Chromium 1.0.0.1.\n'
-            'pick bbb # Conflict-resolved patches from Chromium 1.0.0.1 '
-            'to Chromium 1.0.0.2.\n')
-
-        brockit.Rebase.squash_minor_bumps_from_rebase_plan(path)
-
-        self.assertEqual(
-            path.read_text(),
-            'pick aaa # Conflict-resolved patches from Chromium 1.0.0.0 '
-            'to Chromium 1.0.0.1.\n'
-            'squash bbb # Conflict-resolved patches from Chromium 1.0.0.1 '
-            'to Chromium 1.0.0.2.\n')
-
-    def test_squash_collapses_gnrt_runs(self):
-        """Repeated `gnrt` run commits keep first as pick."""
-        path = self._todo(
-            'pick aaa # [cr148] `gnrt` run for Chromium 1.0.0.1\n'
-            'pick bbb # [cr148] `gnrt` run for Chromium 1.0.0.2\n')
-
-        brockit.Rebase.squash_minor_bumps_from_rebase_plan(path)
-
-        self.assertEqual(
-            path.read_text(),
-            'pick aaa # [cr148] `gnrt` run for Chromium 1.0.0.1\n'
-            'squash bbb # [cr148] `gnrt` run for Chromium 1.0.0.2\n')
-
-    def test_squash_collapses_iwyu_fixes(self):
-        """Repeated `IWYU fixes.` commits keep first as pick."""
-        path = self._todo('pick aaa # [cr148] IWYU fixes.\n'
-                          'pick bbb # [cr148] IWYU fixes.\n')
-
-        brockit.Rebase.squash_minor_bumps_from_rebase_plan(path)
-
-        self.assertEqual(
-            path.read_text(), 'pick aaa # [cr148] IWYU fixes.\n'
-            'squash bbb # [cr148] IWYU fixes.\n')
-
-    def test_squash_reassign_matched_by_hash_moves_above_target(self):
-        """A `reassign!<hash>!` commit is inserted directly above the `pick`
-        for that hash, and that target line becomes `squash`."""
-        path = self._todo('pick aaa # [cr148] Feature A\n'
-                          'pick bbb # [cr148] Feature B\n'
-                          'pick zzz # reassign!bbb! [cr148] Feature B\n')
-
-        brockit.Rebase.squash_minor_bumps_from_rebase_plan(path)
-
-        # `aaa` and `bbb` are both "others"; reassign for `bbb` moves above it
-        # and turns `bbb`'s pick into squash.
-        self.assertEqual(
-            path.read_text(), 'pick aaa # [cr148] Feature A\n'
-            'pick zzz  # reassign!bbb! [cr148] Feature B\n'
-            'squash bbb # [cr148] Feature B\n')
-
-    def test_squash_reassign_matched_by_message_when_hash_mismatch(self):
-        """If the reassign hash doesn't match any pick, fall back to matching
-        a line whose comment ends with the reassign message."""
-        path = self._todo(
-            'pick aaa # [cr148] Feature A\n'
-            'pick bbb # [cr148] Feature B\n'
-            # The hash referenced in the reassign (`xxx`) doesn't appear; the
-            # message `[cr148] Feature B` matches `bbb`'s comment.
-            'pick zzz # reassign!xxx! [cr148] Feature B\n')
-
-        brockit.Rebase.squash_minor_bumps_from_rebase_plan(path)
-
-        self.assertEqual(
-            path.read_text(), 'pick aaa # [cr148] Feature A\n'
-            'pick zzz  # reassign!xxx! [cr148] Feature B\n'
-            'squash bbb # [cr148] Feature B\n')
-
-    def test_squash_orphaned_reassign_is_silently_dropped(self):
-        """A reassign whose hash and message both fail to match is removed."""
-        path = self._todo('pick aaa # [cr148] Feature A\n'
-                          'pick zzz # reassign!xxx! [cr148] Some completely '
-                          'unrelated message\n')
-
-        brockit.Rebase.squash_minor_bumps_from_rebase_plan(path)
-
-        self.assertEqual(path.read_text(), 'pick aaa # [cr148] Feature A\n')
-
-    def test_squash_reassign_strips_empty_suffix_for_message_lookup(self):
-        """An empty reassign carries a trailing ` # empty` marker which must be
-        stripped before the message-based lookup."""
-        path = self._todo(
-            'pick aaa # [cr148] Feature A\n'
-            'pick bbb # [cr148] Feature B\n'
-            'pick zzz # reassign!xxx! [cr148] Feature B # empty\n')
-
-        brockit.Rebase.squash_minor_bumps_from_rebase_plan(path)
-
-        # The ` # empty` suffix is stripped so `[cr148] Feature B` matches
-        # `bbb`'s comment.
-        self.assertEqual(
-            path.read_text(), 'pick aaa # [cr148] Feature A\n'
-            'pick zzz  # reassign!xxx! [cr148] Feature B # empty\n'
-            'squash bbb # [cr148] Feature B\n')
-
-    def test_squash_drops_comments_and_blank_lines(self):
-        """Comment lines and blank lines are not categorised and are absent
-        from the rewritten file."""
-        path = self._todo('# A comment\n'
-                          '\n'
-                          'pick aaa # [cr148] Real commit\n'
-                          '\n'
-                          '# Another comment\n')
-
-        brockit.Rebase.squash_minor_bumps_from_rebase_plan(path)
-
-        self.assertEqual(path.read_text(), 'pick aaa # [cr148] Real commit\n')
-
-    # ----- fix_squash_commit_messages ----------------------------------------
-
-    def test_fix_squash_standard_keeps_last_valid_line(self):
-        """A standard squash commit-message file collapses to just the last
-        non-comment, non-blank line."""
-        path = self._todo('# This is the 1st commit message:\n'
-                          'Update from Chromium 1.0.0.0 to Chromium 1.0.0.1\n'
-                          '\n'
-                          '# This is the commit message #2:\n'
-                          'Update from Chromium 1.0.0.1 to Chromium 1.0.0.2\n')
-
-        brockit.Rebase.fix_squash_commit_messages(path)
-
-        self.assertEqual(path.read_text(),
-                         'Update from Chromium 1.0.0.1 to Chromium 1.0.0.2\n')
-
-    def test_fix_squash_reassign_keeps_following_content(self):
-        """When the first valid line is `reassign!...`, drop it and keep
-        everything from the next *valid* line's index onward in the file."""
-        path = self._todo('# This is the 1st commit message:\n'
-                          'reassign!bbb! [cr148] Feature B\n'
-                          '\n'
-                          '# This is the commit message #2:\n'
-                          '[cr148] Feature B\n'
-                          '\n'
-                          'Original body line.\n')
-
-        brockit.Rebase.fix_squash_commit_messages(path)
-
-        # The cut starts at the next valid line's *index*, not the line
-        # before it -- so the `# This is the commit message #2:` comment is
-        # dropped along with the reassign line. Blanks and content after the
-        # cut point survive.
-        self.assertEqual(path.read_text(), '[cr148] Feature B\n'
-                         '\n'
-                         'Original body line.\n')
-
-    def test_fix_squash_empty_file_exits(self):
-        """A file with only comments and blanks calls `sys.exit`."""
-        path = self._todo('# Only a comment\n\n# And another\n')
-
-        with self.assertRaises(SystemExit):
-            brockit.Rebase.fix_squash_commit_messages(path)
-
-    def test_fix_squash_reassign_with_no_following_content_exits(self):
-        """A reassign squash with no content after the `reassign!` line
-        cannot be collapsed and calls `sys.exit`."""
-        path = self._todo('# This is the 1st commit message:\n'
-                          'reassign!bbb! [cr148] Feature B\n')
-
-        with self.assertRaises(SystemExit):
-            brockit.Rebase.fix_squash_commit_messages(path)
-
-
-@dataclass(frozen=True)
-class _Scenario:
-    """Scenario data returned by `_seed_rebase_scenario` for assertions.
-
-    The three `vNNN` fields are commit hashes on master for the version-bump
-    commits (`Update from Chromium ... to Chromium 1.0.0.N`). `feature` is
-    the Brave-only commit on the feature branch.
-    """
-    branch: str
-    v100: str
-    v101: str
-    v102: str
-    feature: str
-
 
 class RebaseExecuteTest(unittest.TestCase):
     """End-to-end tests for `Rebase.execute()` against a real git tree."""
@@ -376,98 +1081,179 @@ class RebaseExecuteTest(unittest.TestCase):
         self.repo.create_brave_remote()
         self._commit_counter = 0
 
-    # ----- Seeding helpers ---------------------------------------------------
-
-    def _seed_rebase_scenario(self,
-                              branch_name: str = 'feature-branch'
-                              ) -> _Scenario:
-        """Builds a minor-version-bump scenario.
-
-        Layout produced (commits in order):
-            initial -> v100 -> v101  (master)
-                              \\-> feature_commit  (branch_name, on origin)
-                       \\-> v102 (master continues)
-
-        After this returns, the working tree is on `branch_name`.
-        """
+    def _seed_bump_branch(self) -> dict:
+        """Builds a small branch: v100 → v101 (master) → feature → v102
+        (master continues). Returns the relevant hashes."""
         brave = self.repo.brave
         v100 = self.repo.update_brave_version('1.0.0.0')
         v101 = self.repo.update_brave_version('1.0.0.1')
-
-        self.repo._run_git_command(['checkout', '-b', branch_name], brave)
+        self.repo._run_git_command(['checkout', '-b', 'feature-branch'], brave)
         self.repo.write_and_stage_file('feature.txt', 'feature content\n',
                                        brave)
-        feature_commit = self.repo.commit('Add brave-only feature.txt', brave)
+        feature = self.repo.commit('Add brave-only feature.txt', brave)
         self.repo._run_git_command(
-            ['push', '--set-upstream', 'origin', branch_name], brave)
-
+            ['push', '--set-upstream', 'origin', 'feature-branch'], brave)
         self.repo._run_git_command(['checkout', 'master'], brave)
         v102 = self.repo.update_brave_version('1.0.0.2')
-        self.repo._run_git_command(['checkout', branch_name], brave)
-
-        return _Scenario(branch=branch_name,
-                         v100=v100,
-                         v101=v101,
-                         v102=v102,
-                         feature=feature_commit)
-
-    # ----- Small assertion helpers -------------------------------------------
+        self.repo._run_git_command(['checkout', 'feature-branch'], brave)
+        return {'v100': v100, 'v101': v101, 'v102': v102, 'feature': feature}
 
     def _git_log_subjects(self, ref: str = 'HEAD') -> list:
-        """Returns the commit subjects from oldest to newest on `ref`."""
         out = self.repo._run_git_command(
             ['log', '--reverse', '--format=%s', ref], self.repo.brave)
         return out.splitlines() if out else []
 
-    def _git_rev_list_count(self, ref: str = 'HEAD') -> int:
-        """Returns the number of commits reachable from `ref`."""
-        return int(
-            self.repo._run_git_command(['rev-list', '--count', ref],
-                                       self.repo.brave))
-
     def _commit_with_file(self, message: str) -> str:
-        """Creates a commit with a small unique file change.
-
-        Using a real file change (rather than `--allow-empty`) keeps commits
-        from being dropped by `git rebase --empty=drop`, which would obscure
-        what the editor callbacks are actually doing.
-        """
+        """Adds a small unique file and commits it -- a real change so
+        `--empty=drop` doesn't strip it."""
         self._commit_counter += 1
-        counter = self._commit_counter
-        self.repo.write_and_stage_file(f'gen-{counter}.txt',
-                                       f'content {counter}\n', self.repo.brave)
+        self.repo.write_and_stage_file(f'gen-{self._commit_counter}.txt',
+                                       f'content {self._commit_counter}\n',
+                                       self.repo.brave)
         return self.repo.commit(message, self.repo.brave)
 
-    # ----- Tests -------------------------------------------------------------
+    def test_no_flags_is_a_passthrough_rebase(self):
+        """No flags: no plan transform, history preserved on the new
+        target."""
+        scenario = self._seed_bump_branch()
 
-    def test_execute_happy_path_explicit_refs(self):
-        """Explicit `from_ref`/`to_ref` rebases the feature commit onto the
-        new target; the feature branch's subject is preserved."""
-        scenario = self._seed_rebase_scenario()
-
-        brockit.Rebase().execute(from_ref=scenario.v101,
-                                 to_ref=scenario.v102,
+        brockit.Rebase().execute(from_ref=scenario['v101'],
+                                 to_ref=scenario['v102'],
                                  recommit=False,
                                  discard_regen_changes=False,
                                  squash_minor_bumps=False)
 
-        # The feature commit is now on top of v102. We can't compare hashes
-        # (rebase produces new ones) so we compare subjects and parent.
         self.assertEqual(self._git_log_subjects()[-1],
                          'Add brave-only feature.txt')
         parent = self.repo._run_git_command(['rev-parse', 'HEAD^'],
                                             self.repo.brave)
-        self.assertEqual(parent, scenario.v102)
+        self.assertEqual(parent, scenario['v102'])
 
-    def test_execute_resolves_to_ref_upstream_label(self):
+    def test_discard_regen_changes_drops_recyclable(self):
+        scenario = self._seed_bump_branch()
+        self._commit_with_file(
+            'Update patches from Chromium 1.0.0.1 to Chromium 1.0.0.2.')
+        self._commit_with_file('Updated strings for Chromium 1.0.0.2.')
+        self._commit_with_file('[cr148] Another brave-only feature.')
+
+        brockit.Rebase().execute(from_ref=scenario['v101'],
+                                 to_ref=scenario['v102'],
+                                 recommit=False,
+                                 discard_regen_changes=True,
+                                 squash_minor_bumps=False)
+
+        subjects = self._git_log_subjects(scenario['v102'] + '..HEAD')
+        self.assertEqual(subjects, [
+            'Add brave-only feature.txt',
+            '[cr148] Another brave-only feature.',
+        ])
+
+    def test_squash_minor_bumps_collapses_pinned(self):
+        """Squash groups version bumps at the top and keeps the trailing
+        message. Regular commits stay below in arrival order."""
+        scenario = self._seed_bump_branch()
+        self._commit_with_file(
+            'Update from Chromium 1.0.0.2 to Chromium 1.0.0.3')
+        self._commit_with_file('[cr148] Some unrelated feature commit')
+        self._commit_with_file(
+            'Update from Chromium 1.0.0.3 to Chromium 1.0.0.4')
+
+        brockit.Rebase().execute(from_ref=scenario['v101'],
+                                 to_ref=scenario['v102'],
+                                 recommit=False,
+                                 discard_regen_changes=False,
+                                 squash_minor_bumps=True)
+
+        subjects = self._git_log_subjects(scenario['v102'] + '..HEAD')
+        # Version bumps fold into the trailing message; the original
+        # feature commit and the unrelated feature stay in arrival order.
+        self.assertEqual(subjects, [
+            'Update from Chromium 1.0.0.3 to Chromium 1.0.0.4',
+            'Add brave-only feature.txt',
+            '[cr148] Some unrelated feature commit',
+        ])
+
+    def _commit_fixup(self, target: str = 'HEAD') -> str:
+        """Stages a unique gen-N.txt and commits it with `git commit
+        --fixup=<target>`. Mirrors what `git commit --fixup` produces in
+        the wild: a commit whose subject is `fixup! <target subject>`
+        and whose body is empty. Stacking a second call (with the
+        previous fixup as target) yields a `fixup! fixup! <subject>`
+        subject -- the marker-only blocks that broke `MsgBlock.parse`."""
+        self._commit_counter += 1
+        self.repo.write_and_stage_file(f'gen-{self._commit_counter}.txt',
+                                       f'fixup {self._commit_counter}\n',
+                                       self.repo.brave)
+        self.repo._run_git_command(['commit', f'--fixup={target}'],
+                                   self.repo.brave)
+        return self.repo._run_git_command(['rev-parse', 'HEAD'],
+                                          self.repo.brave)
+
+    def test_squash_minor_bumps_with_stacked_fixup_commits(self):
+        """Regression: `git commit --fixup=<conflict-resolved commit>` (and
+        a stacked fixup of that fixup) produces commits whose message is
+        only the autosquash marker -- `fixup! …` and `fixup! fixup! …`
+        with empty bodies. After `--autosquash` chains them next to the
+        pinned target and `squash_minor_bumps` collapses the whole
+        group, git emits those as marker-only `# fixup!` blocks inside
+        the squash editor file. `MsgBlock.parse` used to raise
+        `EditorRecoverableFailure` for them; the rebase must instead
+        complete and keep the latest pinned subject."""
+        scenario = self._seed_bump_branch()
+        first_pinned = self._commit_with_file(
+            'Conflict-resolved patches from Chromium 1.0.0.1 to '
+            'Chromium 1.0.0.2.')
+        inner_fixup = self._commit_fixup(target=first_pinned)
+        self._commit_fixup(target=inner_fixup)
+        self._commit_with_file(
+            'Conflict-resolved patches from Chromium 1.0.0.2 to '
+            'Chromium 1.0.0.3.')
+
+        brockit.Rebase().execute(from_ref=scenario['v101'],
+                                 to_ref=scenario['v102'],
+                                 recommit=False,
+                                 discard_regen_changes=False,
+                                 squash_minor_bumps=True)
+
+        subjects = self._git_log_subjects(scenario['v102'] + '..HEAD')
+        self.assertEqual(subjects, [
+            'Conflict-resolved patches from Chromium 1.0.0.2 to '
+            'Chromium 1.0.0.3.',
+            'Add brave-only feature.txt',
+        ])
+
+    def test_recommit_amends_first_commit(self):
+        """`--recommit` flips the first `pick` to `edit`, then amends and
+        continues -- the rebased commit gets a fresh hash but keeps its
+        subject."""
+        scenario = self._seed_bump_branch()
+        before = self.repo._run_git_command(['rev-parse', 'HEAD'],
+                                            self.repo.brave)
+
+        brockit.Rebase().execute(from_ref=scenario['v101'],
+                                 to_ref=scenario['v102'],
+                                 recommit=True,
+                                 discard_regen_changes=False,
+                                 squash_minor_bumps=False)
+
+        after = self.repo._run_git_command(['rev-parse', 'HEAD'],
+                                           self.repo.brave)
+        self.assertNotEqual(after, before)
+        self.assertEqual(self._git_log_subjects()[-1],
+                         'Add brave-only feature.txt')
+        parent = self.repo._run_git_command(['rev-parse', 'HEAD^'],
+                                            self.repo.brave)
+        self.assertEqual(parent, scenario['v102'])
+
+    def test_resolves_to_ref_upstream_label(self):
         """`to_ref=None` resolves to `@upstream` (`origin/<branch>`); since
         local and origin agree, the rebase is a no-op and the branch HEAD
         remains the feature commit."""
-        scenario = self._seed_rebase_scenario()
+        scenario = self._seed_bump_branch()
         head_before = self.repo._run_git_command(['rev-parse', 'HEAD'],
                                                  self.repo.brave)
 
-        brockit.Rebase().execute(from_ref=scenario.v101,
+        brockit.Rebase().execute(from_ref=scenario['v101'],
                                  to_ref=None,
                                  recommit=False,
                                  discard_regen_changes=False,
@@ -477,9 +1263,9 @@ class RebaseExecuteTest(unittest.TestCase):
                                                 self.repo.brave)
         self.assertEqual(head_after, head_before)
 
-    def test_execute_invalid_from_ref_raises(self):
+    def test_invalid_from_ref_raises(self):
         """A from_ref that resolves to nothing is rejected before git runs."""
-        self._seed_rebase_scenario()
+        self._seed_bump_branch()
 
         with self.assertRaises(brockit.InvalidInputException):
             brockit.Rebase().execute(from_ref='not-a-real-ref-name',
@@ -488,7 +1274,7 @@ class RebaseExecuteTest(unittest.TestCase):
                                      discard_regen_changes=False,
                                      squash_minor_bumps=False)
 
-    def test_execute_no_upstream_raises(self):
+    def test_no_upstream_raises(self):
         """A branch with no `origin/...` upstream cannot use `to_ref=None`."""
         brave = self.repo.brave
         self.repo.update_brave_version('1.0.0.0')
@@ -535,16 +1321,16 @@ class RebaseExecuteTest(unittest.TestCase):
         """Extracts (to_ref, from_ref, branch) from a captured rebase cmd."""
         return cmd[-3], cmd[-2], cmd[-1]
 
-    def test_execute_defaults_from_ref_to_previous_for_same_major(self):
+    def test_defaults_from_ref_to_previous_for_same_major(self):
         """When the target shares HEAD's major version, `from_ref=None`
         resolves via `@previous` -- which is the parent of HEAD's most
         recent package.json change."""
-        scenario = self._seed_rebase_scenario()
+        scenario = self._seed_bump_branch()
 
         ctx, rebase_calls = self._intercept_rebase()
         with ctx:
             brockit.Rebase().execute(from_ref=None,
-                                     to_ref=scenario.v102,
+                                     to_ref=scenario['v102'],
                                      recommit=False,
                                      discard_regen_changes=False,
                                      squash_minor_bumps=False)
@@ -555,11 +1341,11 @@ class RebaseExecuteTest(unittest.TestCase):
         # it back to a full hash for comparison.
         resolved = self.repo._run_git_command(['rev-parse', from_ref],
                                               self.repo.brave)
-        self.assertEqual(to_ref, scenario.v102)
-        self.assertEqual(resolved, scenario.v100)
-        self.assertEqual(branch, scenario.branch)
+        self.assertEqual(to_ref, scenario['v102'])
+        self.assertEqual(resolved, scenario['v100'])
+        self.assertEqual(branch, 'feature-branch')
 
-    def test_execute_defaults_from_ref_to_previous_major_for_new_major(self):
+    def test_defaults_from_ref_to_previous_major_for_new_major(self):
         """When the target is on a different major version, `from_ref=None`
         resolves via `@previous-major` -- which walks back over commits of
         the same major and returns the parent of the major-bump commit."""
@@ -600,91 +1386,7 @@ class RebaseExecuteTest(unittest.TestCase):
         self.assertEqual(resolved, expected_resolved)
         self.assertEqual(branch, 'feature-major')
 
-    def test_execute_recommit_amends_first_rebased_commit(self):
-        """`recommit=True` flips the first `pick` to `edit`, then amends and
-        continues. The rebased commit gets a fresh hash but keeps its
-        subject."""
-        scenario = self._seed_rebase_scenario()
-        before_head = self.repo._run_git_command(['rev-parse', 'HEAD'],
-                                                 self.repo.brave)
-
-        brockit.Rebase().execute(from_ref=scenario.v101,
-                                 to_ref=scenario.v102,
-                                 recommit=True,
-                                 discard_regen_changes=False,
-                                 squash_minor_bumps=False)
-
-        after_head = self.repo._run_git_command(['rev-parse', 'HEAD'],
-                                                self.repo.brave)
-        self.assertNotEqual(after_head, before_head)
-        self.assertEqual(self._git_log_subjects()[-1],
-                         'Add brave-only feature.txt')
-        parent = self.repo._run_git_command(['rev-parse', 'HEAD^'],
-                                            self.repo.brave)
-        self.assertEqual(parent, scenario.v102)
-
-    def test_execute_discard_regen_changes_strips_regen_commits(self):
-        """`discard_regen_changes=True` removes `Update patches ...` and
-        `Updated strings ...` commits from the rebase plan, end-to-end
-        through the real `GIT_SEQUENCE_EDITOR` subprocess callback."""
-        scenario = self._seed_rebase_scenario()
-        # On top of the feature commit, add two regen-like commits and one
-        # genuine feature commit so we can prove only the regen ones are
-        # dropped.
-        self._commit_with_file(
-            'Update patches from Chromium 1.0.0.1 to Chromium 1.0.0.2.')
-        self._commit_with_file(
-            'Updated strings for Chromium 1.0.0.1 to Chromium 1.0.0.2.')
-        self._commit_with_file('[cr148] Another brave-only feature.')
-
-        brockit.Rebase().execute(from_ref=scenario.v101,
-                                 to_ref=scenario.v102,
-                                 recommit=False,
-                                 discard_regen_changes=True,
-                                 squash_minor_bumps=False)
-
-        subjects = self._git_log_subjects(scenario.v102 + '..HEAD')
-        self.assertEqual(subjects, [
-            'Add brave-only feature.txt',
-            '[cr148] Another brave-only feature.',
-        ])
-
-    def test_execute_squash_minor_bumps_collapses_version_bumps(self):
-        """`squash_minor_bumps=True` reorders the plan so version bumps come
-        first and squashes all-but-the-first; the final commit message is the
-        latest version-bump message."""
-        scenario = self._seed_rebase_scenario()
-        # Add an interleaved sequence: another version bump, a feature commit,
-        # then a third version bump.
-        v102_to_v103 = self._commit_with_file(
-            'Update from Chromium 1.0.0.2 to Chromium 1.0.0.3')
-        self._commit_with_file('[cr148] Some unrelated feature commit')
-        self._commit_with_file(
-            'Update from Chromium 1.0.0.3 to Chromium 1.0.0.4')
-
-        # Rebase from v101 (so all the commits above are in scope) onto v102.
-        brockit.Rebase().execute(from_ref=scenario.v101,
-                                 to_ref=scenario.v102,
-                                 recommit=False,
-                                 discard_regen_changes=False,
-                                 squash_minor_bumps=True)
-
-        subjects = self._git_log_subjects(scenario.v102 + '..HEAD')
-        # Version bumps collapse into the first one, taking the LAST message
-        # (per `fix_squash_commit_messages`). The unrelated feature commit
-        # and the original feature commit are kept as "others".
-        self.assertEqual(subjects, [
-            'Update from Chromium 1.0.0.3 to Chromium 1.0.0.4',
-            'Add brave-only feature.txt',
-            '[cr148] Some unrelated feature commit',
-        ])
-        # Defensive: v102_to_v103 is gone (squashed into the merged bump).
-        all_hashes = self.repo._run_git_command(
-            ['rev-list', scenario.v102 + '..HEAD'],
-            self.repo.brave).splitlines()
-        self.assertNotIn(v102_to_v103, all_hashes)
-
-    def test_execute_rebase_failure_raises_invalid_input(self):
+    def test_rebase_failure_raises_invalid_input(self):
         """A genuine git rebase failure (merge conflict on a tracked file) is
         translated into `InvalidInputException`."""
         brave = self.repo.brave
